♿️(search) add a heading for the search results
in the headings hierarchy, it looked like course glimpse headings were
descendants of "Filter courses". This was pretty misleading for screen
reader users that can rely a lot on headings hierarchy to navigate a
page.

Adding a heading for the results block that is the same level as the
filters to prevent this confusion. The heading is maybe not that
necessary in the UI though, so we can just have it for screen readers.