♿️(blogposts) label categories for better screen reader UX
It's not that obvious that the category tags are categories on the
blogpost list page and on the blogpost detail page.

So, at least for screen readers, add a label.

I guess the better thing to do would be to show this label in the UI
too, but it's open to debate. We can argue the "tag" visuals help us
understand enough.