♿️(blogpost mini) change <a> structure for better screen reader UX
the idea is to stop wrapping everything in the <a> so that screen reader
don't announce the whole glimpse as the link. Because this is difficult
to understand.

Adding a link around the cover image + in the title, so that most of the
glimpse is clickable as usually intended on the web on those kind of
components.