♿️(blogpost mini) make title the 1st thing screen readers announce
Since the categories are related to the article, all this
info should appear *after* the article heading in the DOM.

Headings are a way to navigate inside the page when using a screen
reader. I can jump from heading to heading to parse content.
So if I have some content related to the section described by this
heading, that is written *before* the heading, I can easily miss it.

Now I can't miss it.